Add support for udev's hwdb
udev recently gained a hardware database that is intended to replace large
udev rule files such as the one shipped with media-player-info. This should
give a significant (>50%) speed-up in the processing of usb add events.

Add support for converting mpi to hwdb (where applicable), and restrict the
mpi2udev tool to only output the entries that cannot be represented in the hwdb
format. mpi2udev.py now must be called with "hwdb" or "udev" as first argument.

Stop shipping the .hwdb and .rules files in tarballs as they are not static
any more. Which variant to build is now determined by configure (udev >= 196).
This now needs python2 to build, so add check for it to configure.ac.

Co-Authored-By: Martin Pitt <martin.pitt@ubuntu.com>